+++ lib/Elastica/Connection.php
@@ -342,4 +342,20 @@ class Connection extends Param
 
         return $connection;
     }
+
+    /**
+     * @return string User
+     */
+    public function getUsername()
+    {
+        return $this->hasParam('username') ? $this->getParam('username') : null;
+    }
+
+    /**
+     * @return string Password
+     */
+    public function getPassword()
+    {
+        return $this->hasParam('password') ? $this->getParam('password') : null;
+    }
 }